Move hooks into their own files
